Added permission checker.
If an app consists of more than one module, like in this case (the demo, "app" module and the "fileDialog" module)
Then inside of each mdoules manifest file the permissions needed have to be decleared.

Each modules componet (usually the main activity) has to check if these permissions are granted.